requirements.txt
requests~=2.31.0
flask~=3.0.2
pillow~=10.2.0